fix attribute value type

+++ src/main/java/user11681/usersmanual/item/ItemModifiers.java
@@ -9,27 +9,27 @@ public final class ItemModifiers extends Item {
         super(null);
     }
 
-    public static EntityAttributeModifier createAttackSpeedModifier(final int value) {
+    public static EntityAttributeModifier createAttackSpeedModifier(final double value) {
         return createAttackSpeedModifier(value, EntityAttributeModifier.Operation.ADDITION);
     }
 
-    public static EntityAttributeModifier createAttackSpeedModifier(final int value, final EntityAttributeModifier.Operation operation) {
+    public static EntityAttributeModifier createAttackSpeedModifier(final double value, final EntityAttributeModifier.Operation operation) {
         return createAttackSpeedModifier("Weapon modifier", value, operation);
     }
 
-    public static EntityAttributeModifier createAttackSpeedModifier(final String name, final int value, final EntityAttributeModifier.Operation operation) {
+    public static EntityAttributeModifier createAttackSpeedModifier(final String name, final double value, final EntityAttributeModifier.Operation operation) {
         return new EntityAttributeModifier(ATTACK_SPEED_MODIFIER_ID, name, value, operation);
     }
 
-    public static EntityAttributeModifier createAttackDamageModifier(final int value) {
+    public static EntityAttributeModifier createAttackDamageModifier(final double value) {
         return createAttackDamageModifier(value, EntityAttributeModifier.Operation.ADDITION);
     }
 
-    public static EntityAttributeModifier createAttackDamageModifier(final int value, final EntityAttributeModifier.Operation operation) {
+    public static EntityAttributeModifier createAttackDamageModifier(final double value, final EntityAttributeModifier.Operation operation) {
         return createAttackDamageModifier("Weapon modifier", value, operation);
     }
 
-    public static EntityAttributeModifier createAttackDamageModifier(final String name, final int value, final EntityAttributeModifier.Operation operation) {
+    public static EntityAttributeModifier createAttackDamageModifier(final String name, final double value, final EntityAttributeModifier.Operation operation) {
         return new EntityAttributeModifier(ATTACK_DAMAGE_MODIFIER_ID, name, value, operation);
     }
 }
